Restructure biome::user module
Moves the entire user module under user::store, to match the expected
future structure of peer modules. (Currently, this user module is so
simple that it only has a store, but if this module survives, it may
also grow to match its peer modules.)

Moves all code requiring diesel to user::diesel. This makes room for
other store backends (such as memory, etc.).

Moves all PostgreSQL-specific code to user::diesel::postgres, to make
room for other database backend implementations.
